Use class_prepared signal for loading fixtures
If we do all of our loading in AppConfig.ready, then it runs before the
normal Django checks framework. This means that errors in our app models
throw a stack trace instead of a more detailed message if there is a
problem with our schema. By moving our schema loading to a signal, we
can insure that the checks framework runs first.

It's possible some of the other checks for ImproperlyConfigured could be
 moved to the checks framework in a future PR.