Move loader_class to fix autoreload signal
We need to register autoreload signal in our AppConfig.ready so that it
registers soon enough, but we also need to be able to reference the EXT_
that we want to use. We move the loader_class config to our ready()
method so that things load in the correct order without errors.